feat(structure): refactored structure of project
