+++ README.md
@@ -1,1 +1,26 @@
-# voxel-slam-utils
+# Voxel-SLAM-utils
+
+Contains utility scripts for interacting with [voxel-based](https://github.com/prime-slam/voxel-slam) pipeline.
+
+# Installation
+
+To use this library you need to download this repository and install necessary dependencies.
+How to do this:
+
+1. Download and install Python 3.10 from the [official website](https://www.python.org/downloads/).
+2. Clone this repository
+    ```bash
+    git clone https://github.com/prime-slam/voxel-slam.git
+    ```
+3. Create and activate virtual environment
+    ```bash
+    python3 -m venv venv && source venv/bin/activate
+    ```
+4. Install dependencies
+    ```bash
+    pip install -r requirements.txt
+    ```
+
+# Usage
+
+All folders contain the necessary instructions to execute the scripts in README.md files.
