+++ hilti/initial/prepare_clouds.py
