+++ hilti/orb-slam/prepare_poses.py
