+++ hilti/utils/__init__.py
@@ -1,0 +1,4 @@
+import hilti.utils.utils as utils_module
+from hilti.utils.utils import *
+
+__all__ = utils_module.__all__
