+++ hilti/visualize.py
@@ -1,38 +1,10 @@
-import mrob
-import numpy as np
 import open3d as o3d
 
 import argparse
 import os
 import random
 
-
-def get_imu():
-    # TODO: change values for specific IMU matrix
-    R = mrob.geometry.quat_to_so3([0.7071068, -0.7071068, 0, 0])
-    T = np.eye(4)
-    T[:3, :3] = R
-    T[:3, 3] = [-0.001, -0.00855, 0.055]
-
-    return T
-
-
-def get_sorted(directory):
-    with os.scandir(directory) as entries:
-        sorted_entries = sorted(entries, key=lambda entry: entry.name)
-        sorted_items = [entry.name for entry in sorted_entries]
-    return sorted_items
-
-
-def read_pose(path: str) -> np.ndarray:
-    pose = np.eye(4)
-    with open(path) as file:
-        lines = file.readlines()
-        for ind, line in enumerate(lines):
-            pose[ind] = list(map(float, line.replace("\n", "").split(" ")))
-
-    return pose
-
+from hilti.utils.utils import get_lidar_to_imu, get_sorted, read_pose
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(prog="VisualizePoses")
@@ -44,16 +16,16 @@ if __name__ == "__main__":
 
     point_cloud = o3d.geometry.PointCloud(o3d.utility.Vector3dVector())
     point_clouds = get_sorted(args.dataset)[args.start : args.end]
-    imu = get_imu()
+    lidar_to_imu = get_lidar_to_imu()
     random.seed(42)
 
     for point_cloud_name in point_clouds:
         pose_name = point_cloud_name[: point_cloud_name.index(".") + 2] + ".txt"
         try:
-            pose = read_pose(os.path.join(args.poses, pose_name)) @ imu
+            pose = read_pose(os.path.join(args.poses, pose_name)) @ lidar_to_imu
         except Exception as e:
-            print(f"{pose_name} is missing due to {e}, using IMU")
-            pose = imu
+            print(f"{pose_name} is missing due to {e}, using lidar-to-imu")
+            pose = lidar_to_imu
 
         cloud = o3d.io.read_point_cloud(
             os.path.join(args.dataset, point_cloud_name)
