+++ requirements.txt
@@ -1,3 +1,4 @@
 mrob==0.0.14
 numpy==1.26.3
 open3d==0.18.0
+colour==0.1.5
